requirements.txt
streamlit>=1.30
pandas>=2.0
python-dotenv>=1.0

